update streaming options
